+++ apps/walktest/walkcmd.cpp
@@ -2452,10 +2452,10 @@ bool CommandHandler (const char *cmd, const char *arg)
         
         int width = QInt(bbox.MaxX() - bbox.MinX());
         int height = QInt(bbox.MaxY() - bbox.MinY());
-        Sys->views[0]->SetRectangle(bbox.MinX(), bbox.MinY(), width / 2, height);
+        Sys->views[0]->SetRectangle((int)bbox.MinX(), (int)bbox.MinY(), width / 2, height);
         Sys->views[0]->GetCamera()->SetPerspectiveCenter(bbox.MinX() + (width / 4),
                                                         bbox.MinY() + (height / 2));
-        Sys->views[1]->SetRectangle(bbox.MinX() + (width / 2), bbox.MinY(), 
+        Sys->views[1]->SetRectangle((int)bbox.MinX() + (width / 2), (int)bbox.MinY(), 
                                     width / 2, height);
         Sys->views[1]->GetCamera()->SetPerspectiveCenter(bbox.MinX() + (3 * width / 4),
                                                         bbox.MinY() + (height / 2));
@@ -2473,7 +2473,7 @@ bool CommandHandler (const char *cmd, const char *arg)
 
         int width = QInt(bbox2.MaxX() - bbox1.MinX());
         int height = QInt(bbox1.MaxY() - bbox1.MinY());
-        Sys->view->SetRectangle(bbox1.MinX(), bbox1.MinY(), width, height);
+        Sys->view->SetRectangle((int)bbox1.MinX(), (int)bbox1.MinY(), width, height);
         Sys->view->GetCamera()->SetPerspectiveCenter(bbox1.MinX() + (width / 2), 
                                                     bbox2.MinY() + (height / 2));
         Sys->split = -1;
